requirements.txt
dialogflow==0.6.0
python-telegram-bot==11.1.0
googleapis-common-protos==1.6.0
python-dotenv==0.10.3
vk-api==11.5.0
